+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.6] - 07/10/25
+
+### Added
+
+- Added set_default_checkers method to Backgammon Game class for setting the initial position of the checkers at the beginning of the game.
+
+## [0.1.5] - 07/10/25
+
+### Changed
+
+- Justificacion.md created and all the information in README.md was transferred as I mistakenly did that part the other way around.
+
 ## [0.1.4] - 29/09/25
 
 ### Added